Rename `ActionView::Base#run`
There was a recent change by @tenderlove to Action view which introduced
`ActionView::Base#run` [1].

We ran into an issue with our application because one of the core
concepts in our domain model is a `Run` which is exposed in most of our
views as a helper method, which now conflicts with this new method.

In order to save us having to rewrite a substantial part of our app I
was wondering if we might be able to rename this method instead to
something like `#call` or perhaps `#exec`?

[1] https://github.com/rails/rails/commit/c740ebdaf5